fixed issue: being closed when the inside of dropdown field is clicked while it is being scrolled.
in IOS, I clicked the dropdown's scroll view to stop scrolling, and I expected that it would stop but the dropdown field was closed.

originally, because the dropdown field had been wrapped with a GestureDetector as a child, the gesture action would be applied to that child too, where it shouldn't happen.

Instead I separated and placed them in a Stack, and it was fixed.

No side issue was found.