Fix: Bug which prevented the onChanged to be invoked after first invocation
In the custom_dropdown.dart, the didUpdateWidget method changes the selectedItemNotifier
and selectedItemsNotifier to a new instance when the widget configuration is inconsistent.

This causes two problems, first if the end user supplied the controller then it will be overriden. Secondly, new instance is not
required, and even if created should have attached the listeners to the the new instance.

This commit solves the problem by updating the earlier instance of the notifier to the new value.
Also, the two examples have been changed to use the stateful widget, which solves the current problem.